Extract classic raw data lowering
